+++ .idea/.idea.LLL/.idea/contentModel.xml
@@ -8,7 +8,6 @@
       <e p="Task3.API" t="IncludeRecursive">
         <e p="Controllers" t="Include">
           <e p="MaterialController.cs" t="Include" />
-          <e p="MaterialVersionController.cs" t="Include" />
         </e>
         <e p="Mapping" t="Include">
           <e p="MappingProfile.cs" t="Include" />
@@ -24,6 +23,10 @@
           <e p="Material_06.jpeg" t="Include" />
           <e p="Material_07.jpeg" t="Include" />
           <e p="Material_08.jpeg" t="Include" />
+          <e p="Material_09.jpeg" t="Include" />
+          <e p="Material_10.jpeg" t="Include" />
+          <e p="Material_11.jpeg" t="Include" />
+          <e p="Material_12.jpeg" t="Include" />
           <e p="m1V1.jpeg" t="Include" />
           <e p="m1V2.jpeg" t="Include" />
           <e p="m1V4.jpeg" t="Include" />
@@ -33,9 +36,9 @@
           <e p="launchSettings.json" t="Include" />
         </e>
         <e p="Resources" t="Include">
-          <e p="MaterialResources.cs" t="Include" />
-          <e p="MaterialVersionResourses.cs" t="Include" />
-          <e p="NewMaterialAddResources.cs" t="Include" />
+          <e p="MaterialDownloadResultDto.cs" t="Include" />
+          <e p="MaterialResultDto.cs" t="Include" />
+          <e p="MaterialVersionResultDto.cs" t="Include" />
         </e>
         <e p="Startup.cs" t="Include" />
         <e p="Task3.API.csproj" t="IncludeRecursive" />
@@ -56,7 +59,6 @@
       </e>
       <e p="Task3.BLL" t="IncludeRecursive">
         <e p="MaterialService.cs" t="Include" />
-        <e p="MaterialVersionService.cs" t="Include" />
         <e p="Task3.BLL.csproj" t="IncludeRecursive" />
         <e p="bin" t="ExcludeRecursive" />
         <e p="obj" t="ExcludeRecursive">
@@ -86,7 +88,6 @@
         </e>
         <e p="Services" t="Include">
           <e p="IMaterialServices.cs" t="Include" />
-          <e p="IMaterialVersionServices.cs" t="Include" />
         </e>
         <e p="Task3.Core.csproj" t="IncludeRecursive" />
         <e p="bin" t="ExcludeRecursive" />
