+++ Task3.API/Controllers/MaterialController.cs
@@ -2,8 +2,10 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using AutoMapper;
+using FluentValidation.Internal;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Newtonsoft.Json.Serialization;
 using Task3.API.Resources;
 using Task3.API.Validations;
 using Task3.Core.DTO;
@@ -25,8 +27,8 @@ namespace Task3.API.Controllers
             this._materialService = materialService;
         }
 
-        [HttpGet("")]
-        public async Task<ActionResult<IEnumerable<Material>>> GetMaterialsByDate()
+        [HttpGet("/api/materials/date")]
+        public async Task<ActionResult<IEnumerable<MaterialResultDto>>> GetMaterialsByDate()
         {
             var materials = await _materialService.GetFilterMaterialsByDate();
             var materialResultDto = _mapper.
@@ -34,27 +36,61 @@ namespace Task3.API.Controllers
             return Ok(materialResultDto);
         }
 
-        [HttpPost("")]
-        public async Task<MaterialResultDto> UploadedMaterial([FromForm] UploadMaterialDTO uploadMaterialForm)
+        [HttpPost("/api/material/upload")]
+        public async Task<ActionResult<MaterialResultDto>> UploadedMaterial([FromForm] UploadMaterialDTO uploadMaterialForm)
         {
             var validator = new SaveMaterialValidator();
             var validationResult = await validator.ValidateAsync(uploadMaterialForm);
-            /*if (!validationResult.IsValid)
-                return BadRequest(validationResult.Errors);*/
+            if (!validationResult.IsValid)
+                return BadRequest("Not valid data");
             var material = await _materialService.UploadNewMaterial(uploadMaterialForm);
             var materialResultDto = _mapper.
                 Map<Material, MaterialResultDto>(material);
-            //return Ok(material);
-            return materialResultDto;
+            return Ok(materialResultDto);
         }
 
-        [Route("GetMat")]
+        [Route("/api/material/{mId}/download")]
         [HttpPost]
         public async Task<ActionResult> DownloadMaterial(Guid mId)
         {
             var fileData = await _materialService.GetDtoForDownloadMaterialAsync(mId);
-            byte[] mas = System.IO.File.ReadAllBytes(fileData.filePath);
-            return File(mas, fileData.fileType, fileData.fileName);
+            return File(fileData.mas, fileData.fileType, fileData.fileName);
+        }
+        
+        [Route("/api/{mId}/versions/date/")]
+        [HttpGet]
+        public async Task<ActionResult<IEnumerable<MaterialVersionResultDto>>>  GetVersionsOrdergniByDate(Guid mId)
+        {
+            var versions = await  _materialService.FilterVersionsByDate(mId);
+            var materialVersionResultDto = _mapper.
+                Map<IEnumerable<MaterialVersion>, IEnumerable<MaterialVersionResultDto>>(versions);
+            return Ok(materialVersionResultDto);
+        }
+        
+        [Route("/api/{mId}/versions/syze/")]
+        [HttpGet]
+        public async Task<ActionResult<IEnumerable<MaterialVersionResultDto>>>  GetVersionsOrdergniBySyze(Guid mId)
+        {
+            var versions = await  _materialService.FilterVersionsBySize(mId);
+            var materialVersionResultDto = _mapper.
+                Map<IEnumerable<MaterialVersion>, IEnumerable<MaterialVersionResultDto>>(versions);
+            return Ok(materialVersionResultDto);
+        }
+
+        [Route("/api/version/upload/")]
+        [HttpPost]
+        public async Task<ActionResult> UploadNewVersionOfMaterial([FromForm] UploadMaterialVersionDTO materialVersionform)
+        {
+            var versionOfMaterial = await _materialService.UploadNewMaterialVersion(materialVersionform);
+            return Ok(versionOfMaterial);
+        }
+        
+        [Route("/api/version/{vId}/download")]
+        [HttpPost]
+        public async Task<ActionResult> DownloadVersionOfMaterial(Guid vId)
+        {
+            var fileData = await _materialService.GetMaterialVersionFile(vId);
+            return File(fileData.mas, fileData.fileType, fileData.fileName);
         }
 
     }
