+++ Task3.API/Resources/MaterialDownloadResultDto.cs
@@ -1,0 +1,7 @@
+namespace Task3.API.Resources
+{
+    public class MaterialDownloadResultDto
+    {
+        
+    }
+}
