--- Task3.API/Resources/MaterialVersionResourses.cs
@@ -1,12 +1,0 @@
-using System;
-
-namespace Task3.API.Resources
-{
-    public class MaterialVersionResources
-    {
-        public Guid Id { get; set; }
-        public string Name { get; set; }
-        public MaterialResultDto Material { get; set; }
-        
-    }
-}
