--- Task3.API/Resources/NewMaterialAddResources.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace Task3.API.Resources
-{
-    public class NewMaterialAddResources
-    {
-        private Guid Id { get; set; }
-        public string MaterialName { get; set; }
-        public int MaterialId { get; set; }
-    }
-}
